Consolidate background thread handling
